reorganize pages and components folders
