+++ src/components/NavBar/NavBar.spec.js
