+++ src/components/SearchInput/SearchInput.spec.js
