+++ src/components/SearchInput/SearchInput.stories.js
