+++ src/pages/DetailPage/DetailPage.spec.js
