+++ src/pages/DetailPage/DetailPage.stories.js
