+++ src/pages/SearchPage/SearchPage.spec.js
